Rename log_output_utils.go to parse_command_output.go and Move output parsing functions from helpers.go to parse_command_output.go
